TIP-872: Move product classes to Akeneo\Enrichment
