+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/controllers.yml
@@ -15,3 +15,55 @@ services:
             - '@doctrine.orm.default_entity_manager'
             - '@pim_comment.remover.comment'
             - '%pim_comment.entity.comment.class%'
+
+    ### External API
+    pim_api.controller.product:
+        class: 'Akeneo\Pim\Enrichment\Bundle\Controller\ExternalApi\ProductController'
+        arguments:
+            - '@pim_catalog.query.product_query_builder_search_after_size_factory'
+            - '@pim_serializer'
+            - '@pim_api.repository.channel'
+            - '@pim_api.checker.query_parameters_product'
+            - '@pim_api.repository.attribute'
+            - '@pim_api.repository.product'
+            - '@pim_api.pagination.offset_hal_paginator'
+            - '@pim_api.pagination.search_after_hal_paginator'
+            - '@pim_api.pagination.parameter_validator'
+            - '@pim_catalog.validator.product'
+            - '@pim_catalog.builder.product'
+            - '@pim_catalog.remover.product'
+            - '@pim_catalog.updater.product'
+            - '@pim_catalog.saver.product'
+            - '@router'
+            - '@pim_catalog.comparator.filter.product'
+            - '@pim_api.stream.product_partial_update_stream'
+            - '@pim_api.security.primary_key_encrypter'
+            - '@pim_catalog.query.product_query_builder_from_size_factory'
+            - '@pim_catalog.builder.product'
+            - '@pim_api.filter.product_attribute_filter'
+            - '@pim_catalog.entity_with_family_variant.add_parent_to_product'
+            - '%pim_api.configuration%'
+
+
+    pim_api.controller.product_model:
+        class: 'Akeneo\Pim\Enrichment\Bundle\Controller\ExternalApi\ProductModelController'
+        arguments:
+            - '@pim_catalog.query.product_model_query_builder_factory'
+            - '@pim_catalog.query.product_model_query_builder_from_size_factory'
+            - '@pim_catalog.query.product_model_query_builder_search_after_size_factory'
+            - '@pim_serializer'
+            - '@pim_api.repository.channel'
+            - '@pim_api.checker.query_parameters_product_model'
+            - '@pim_api.pagination.parameter_validator'
+            - '@pim_api.pagination.offset_hal_paginator'
+            - '@pim_api.pagination.search_after_hal_paginator'
+            - '@pim_api.security.primary_key_encrypter'
+            - '@pim_api.updater.product_model'
+            - '@pim_catalog.factory.product_model'
+            - '@pim_catalog.saver.product_model'
+            - '@router'
+            - '@pim_catalog.validator.product_model'
+            - '@pim_api.filter.product_model_attribute_filter'
+            - '@pim_catalog.repository.product_model'
+            - '@pim_api.stream.product_model_partial_update_stream'
+            - '%pim_api.configuration%'
