+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/normalizers.yml
@@ -3,3 +3,12 @@ services:
         class: 'Akeneo\Pim\Enrichment\Component\Comment\Normalizer\Standard\CommentNormalizer'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
+
+    pim_api.normalizer.product:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Normalizer\ExternalApi\ProductNormalizer'
+        arguments:
+            - '@pim_catalog.normalizer.standard.product'
+            - '@pim_api.repository.attribute'
+            - '@router'
+        tags:
+            - { name: pim_serializer.normalizer, priority: 90 }
