+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/routing/product.yml
@@ -4,3 +4,34 @@ pim_pdf_generator_download_product_pdf:
     requirements:
         id: '[0-9a-f]+'
     methods: [GET]
+
+### External Api
+pim_api_product_list:
+    path: /products
+    defaults: { _controller: pim_api.controller.product:listAction, _format: json }
+    methods: [GET]
+
+pim_api_product_get:
+    path: /products/{code}
+    defaults: { _controller: pim_api.controller.product:getAction, _format: json }
+    methods: [GET]
+
+pim_api_product_create:
+    path: /products
+    defaults: { _controller: pim_api.controller.product:createAction, _format: json }
+    methods: [POST]
+
+pim_api_product_partial_update:
+    path: /products/{code}
+    defaults: { _controller: pim_api.controller.product:partialUpdateAction, _format: json }
+    methods: [PATCH]
+
+pim_api_product_partial_update_list:
+    path: /products
+    defaults: { _controller: pim_api.controller.product:partialUpdateListAction, _format: json }
+    methods: [PATCH]
+
+pim_api_product_delete:
+    path: /products/{code}
+    defaults: { _controller: pim_api.controller.product:deleteAction, _format: json }
+    methods: [DELETE]
