+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/routing/product_model.yml
