+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/updaters.yml
@@ -467,3 +467,8 @@ services:
             - ['pim_catalog_multiselect']
         tags:
             - { name: 'pim_catalog.updater.remover' }
+
+    pim_api.updater.product_model:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\ExternalApi\Updater\ProductModelUpdater'
+        arguments:
+            - '@pim_catalog.updater.product_model'
