+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/controllers.yml
@@ -1,8 +1,6 @@
 parameters:
     pim_api.controller.category.class: Akeneo\Tool\Bundle\ApiBundle\Controller\CategoryController
     pim_api.controller.token.class: Akeneo\Tool\Bundle\ApiBundle\Controller\TokenController
-    pim_api.controller.product.class: Akeneo\Tool\Bundle\ApiBundle\Controller\ProductController
-    pim_api.controller.product_model.class: Akeneo\Tool\Bundle\ApiBundle\Controller\ProductModelController
     pim_api.controller.root_endpoint.class: Akeneo\Tool\Bundle\ApiBundle\Controller\RootEndpointController
     pim_api.controller.media_file.class: Akeneo\Tool\Bundle\ApiBundle\Controller\MediaFileController
     pim_api.controller.currency.class: Akeneo\Tool\Bundle\ApiBundle\Controller\CurrencyController
@@ -28,56 +26,6 @@ services:
         class: '%pim_api.controller.token.class%'
         arguments:
             - '@fos_oauth_server.server'
-
-    pim_api.controller.product:
-        class: '%pim_api.controller.product.class%'
-        arguments:
-            - '@pim_catalog.query.product_query_builder_search_after_size_factory'
-            - '@pim_serializer'
-            - '@pim_api.repository.channel'
-            - '@pim_api.checker.query_parameters_product'
-            - '@pim_api.repository.attribute'
-            - '@pim_api.repository.product'
-            - '@pim_api.pagination.offset_hal_paginator'
-            - '@pim_api.pagination.search_after_hal_paginator'
-            - '@pim_api.pagination.parameter_validator'
-            - '@pim_catalog.validator.product'
-            - '@pim_catalog.builder.product'
-            - '@pim_catalog.remover.product'
-            - '@pim_catalog.updater.product'
-            - '@pim_catalog.saver.product'
-            - '@router'
-            - '@pim_catalog.comparator.filter.product'
-            - '@pim_api.stream.product_partial_update_stream'
-            - '@pim_api.security.primary_key_encrypter'
-            - '@pim_catalog.query.product_query_builder_from_size_factory'
-            - '@pim_catalog.builder.product'
-            - '@pim_api.filter.product_attribute_filter'
-            - '@pim_catalog.entity_with_family_variant.add_parent_to_product'
-            - '%pim_api.configuration%'
-
-    pim_api.controller.product_model:
-        class: '%pim_api.controller.product_model.class%'
-        arguments:
-            - '@pim_catalog.query.product_model_query_builder_factory'
-            - '@pim_catalog.query.product_model_query_builder_from_size_factory'
-            - '@pim_catalog.query.product_model_query_builder_search_after_size_factory'
-            - '@pim_serializer'
-            - '@pim_api.repository.channel'
-            - '@pim_api.checker.query_parameters_product_model'
-            - '@pim_api.pagination.parameter_validator'
-            - '@pim_api.pagination.offset_hal_paginator'
-            - '@pim_api.pagination.search_after_hal_paginator'
-            - '@pim_api.security.primary_key_encrypter'
-            - '@pim_api.updater.product_model'
-            - '@pim_catalog.factory.product_model'
-            - '@pim_catalog.saver.product_model'
-            - '@router'
-            - '@pim_catalog.validator.product_model'
-            - '@pim_api.filter.product_model_attribute_filter'
-            - '@pim_catalog.repository.product_model'
-            - '@pim_api.stream.product_model_partial_update_stream'
-            - '%pim_api.configuration%'
 
     pim_api.controller.root_endpoint:
         class: '%pim_api.controller.root_endpoint.class%'
