+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/normalizers.yml
@@ -2,7 +2,6 @@ parameters:
     pim_api.normalizer.exception.documented.class: Akeneo\Tool\Component\Api\Normalizer\Exception\DocumentedNormalizer
     pim_api.normalizer.exception.violation.class: Akeneo\Tool\Component\Api\Normalizer\Exception\ViolationNormalizer
     pim_api.normalizer.collection.class: Akeneo\Tool\Component\Api\Normalizer\CollectionNormalizer
-    pim_api.normalizer.product.class: Akeneo\Tool\Component\Api\Normalizer\ProductNormalizer
     pim_api.normalizer.product_model.class: Pim\Component\Api\Normalizer\ProductModelNormalizer
     pim_api.normalizer.category.class: Akeneo\Tool\Component\Api\Normalizer\CategoryNormalizer
     pim_api.normalizer.attribute_group.class: Akeneo\Pim\Structure\Component\Normalizer\ExternalApi\AttributeGroupNormalizer
@@ -34,15 +33,6 @@ services:
 
     pim_api.normalizer.collection:
         class: '%pim_api.normalizer.collection.class%'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_api.normalizer.product:
-        class: '%pim_api.normalizer.product.class%'
-        arguments:
-            - '@pim_catalog.normalizer.standard.product'
-            - '@pim_api.repository.attribute'
-            - '@router'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
