--- src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/updaters.yml
@@ -1,8 +1,0 @@
-parameters:
-    pim_api.updater.product_model.class: Akeneo\Tool\Component\Api\Updater\ProductModelUpdater
-
-services:
-    pim_api.updater.product_model:
-        class: '%pim_api.updater.product_model.class%'
-        arguments:
-            - '@pim_catalog.updater.product_model'
